+++ EssentialFeedTests/Feed Presentation/FeedImagePresenter.swift
@@ -1,0 +1,65 @@
+//
+//  FeedImagePresenter.swift
+//  EssentialFeedTests
+//
+//  Created by Julius on 01/08/2022.
+//
+
+import Foundation
+
+public struct FeedImageViewModel<Image> {
+    public let description: String?
+    public let location: String?
+    public let image: Image?
+    public let isLoading: Bool
+    public let shouldRetry: Bool
+
+    public var hasLocation: Bool {
+        return location != nil
+    }
+}
+
+public protocol FeedImageView {
+    associatedtype Image
+    func display(_ model: FeedImageViewModel<Image>)
+}
+
+public final class FeedImagePresenter<View: FeedImageView, Image> where View.Image == Image {
+    private let view: View
+    private let imageTransformer: (Data) -> Image?
+
+
+    public init(view: View, imageTransformer: @escaping (Data) -> Image?) {
+        self.view = view
+        self.imageTransformer = imageTransformer
+
+    }
+    
+    public func didStartLoadingImageData(for model: FeedItem) {
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: nil,
+            isLoading: true,
+            shouldRetry: false))
+    }
+    
+    public func didFinishLoadingImageData(with data: Data, for model: FeedItem) {
+        let image = imageTransformer(data)
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: image,
+            isLoading: false,
+            shouldRetry: image == nil))
+    }
+    
+    public func didFinishLoadingImageData(with error: Error, for model: FeedItem) {
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: nil,
+            isLoading: false,
+            shouldRetry: true))
+    }
+}
