+++ EssentialFeedTests/Feed Presentation/FeedImagePresenterTests.swift
@@ -8,63 +8,6 @@
 import XCTest
 import EssentialFeed
 
-struct FeedImageViewModel<Image> {
-    let description: String?
-    let location: String?
-    let image: Image?
-    let isLoading: Bool
-    let shouldRetry: Bool
-
-    var hasLocation: Bool {
-        return location != nil
-    }
-}
-
-protocol FeedImageView {
-    associatedtype Image
-    func display(_ model: FeedImageViewModel<Image>)
-}
-
-final class FeedImagePresenter<View: FeedImageView, Image> where View.Image == Image {
-    private let view: View
-    private let imageTransformer: (Data) -> Image?
-
-
-    init(view: View, imageTransformer: @escaping (Data) -> Image?) {
-        self.view = view
-        self.imageTransformer = imageTransformer
-
-    }
-    
-    func didStartLoadingImageData(for model: FeedItem) {
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: nil,
-            isLoading: true,
-            shouldRetry: false))
-    }
-    
-    func didFinishLoadingImageData(with data: Data, for model: FeedItem) {
-        let image = imageTransformer(data)
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: image,
-            isLoading: false,
-            shouldRetry: image == nil))
-    }
-    
-    func didFinishLoadingImageData(with error: Error, for model: FeedItem) {
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: nil,
-            isLoading: false,
-            shouldRetry: true))
-    }
-}
-
 class FeedImagePresenterTests: XCTestCase {
 
     func test_init_doesNotSendMessagesToView() {
